+++ mpv.net/WPF/EverythingWindow.xaml.cs
@@ -142,9 +142,7 @@ namespace mpvnet
                         Everything_GetResultFullPathName(i, sb, (uint)sb.Capacity);
                         string ext = sb.ToString().Ext();
 
-                        if (App.AudioTypes.Contains(ext) || App.VideoTypes.Contains(ext)
-                            || App.ImageTypes.Contains(ext))
-
+                        if (App.AudioTypes.Contains(ext) || App.VideoTypes.Contains(ext) || App.ImageTypes.Contains(ext))
                             items.Add(sb.ToString());
 
                         if (items.Count > 100)