+++ Changelog.md
@@ -1,12 +1,14 @@
 
-5.4.8.7 Beta (202?-??-??)
+5.4.8.7 Beta (2021-03-??)
 =========================
 
 - History feature can be configured to ingore defined strings:
   script-opt = history-discard=value1;value2
 - Web stream audio and subtitle track selection.
-- On Windows 10 1903 and later the default code page was changed to UTF8.
+- On Windows 10 1903 and later the default code page was changed to UTF-8.
 - Support of --version command.
+- File associations and auto-load-folder can be customized with
+  video-file-extensions, audio-file-extensions and image-file-extensions
 
 
 5.4.8.6 Beta (2020-12-24)